review - adjust to fabric workloads
